bot/requirements.txt
# Enviroment
python-environ~=0.4.54

# Telegram
aiogram~=3.20.0  # general aiogram
aiogram-dialog[tools]~=2.3.1  # easy dialogs for aiogram
# uvloop  # fast, drop-in replacement of the built-in asyncio event loop. Windows not supported (uncomment if you are on Linux)
ujson~=5.10.0  # ultra fast JSON encoder and decoder written in pure C
aiohttp[speedups]~=3.10.11
# aiohttp[speedups]==3.9.5 # for windows use this instead (bcs conflicts with aiodns)
# cchardet # (aiohttp speedup) high speed universal character encoding detector
aiodns~=3.4.0 # (aiohttp speedup) provides a simple way for doing asynchronous DNS resolutions

# Cache
redis[hiredis]~=5.2.1

# Logging
structlog~=25.3.0
colorama~=0.4.6

# Localization
fluent.runtime~=0.4.0
fluent.syntax~=0.19.0

# Templates (.docx)
docxtpl~=0.20.0
jsonschema~=4.23.0

###############################################################
###     Some useful async libraries to use with aiogram     ###
### for more: https://github.com/timofurrer/awesome-asyncio ###
##############################################################

# aiosqlite  # Async work with SQLite
# asyncpg  # Async work with PostgreSQL
# motor  # Async work with MongoDB
# pyzmq  # Async work with ZeroMQ
# piccolo  # Async ORM
# aql  # Async query builder (beta)
# httpx  # Async http client
# aiofiles  # Async work with files
# aioamqp  # Async work with AMQP
# aiocache  # Async work with cache system
# aiopath  # Async pathlib